Fix dbkey setup
Fixes vsnp* tool test failures (`Input staging problem: Invalid dbkey
specified`) and
https://sentry.galaxyproject.org/sentry/test/issues/44049/.
Minor bug introduced in
https://github.com/galaxyproject/galaxy/pull/10667/files.
